Revert "Merge branch 'master' of https://github.com/MCDelta/TuxWeapons"
This reverts commit a71fc5dd01f1d44a0bef1d6597b99898eda2c6d5, reversing
changes made to 2d7daaa1503c6b5cae22a7a1d0e12147e6eccf6c.

+++ common/mcdelta/tuxweapons/event/EventFOVModifier.java
@@ -12,28 +12,27 @@ import cpw.mods.fml.relauncher.SideOnly;
 
 public class EventFOVModifier
 {
-    @ForgeSubscribe
-    @SideOnly(Side.CLIENT)
-    public void fovUpdate(FOVUpdateEvent event)
-    {
-        EntityPlayer player = event.entity;
-
-        if (player.isUsingItem()
-                && ((player.getItemInUse().getItem() instanceof ItemDeltaBow) || (player.getItemInUse().getItem() instanceof ItemSpear)
-                        || (player.getItemInUse().getItem() instanceof ItemHammer) || (player.getItemInUse().getItem() instanceof ItemGrappHook)))
-        {
-            int i = player.getItemInUseDuration();
-            float f1 = i / 20.0F;
-
-            if (f1 > 1.0F)
-            {
-                f1 = 1.0F;
-            } else
-            {
-                f1 *= f1;
-            }
-
-            event.newfov *= 1.0F - (f1 * 0.15F);
-        }
-    }
+     @ForgeSubscribe
+     @SideOnly (Side.CLIENT)
+     public void fovUpdate (FOVUpdateEvent event)
+     {
+          EntityPlayer player = event.entity;
+          
+          if (player.isUsingItem() && (player.getItemInUse().getItem() instanceof ItemDeltaBow || player.getItemInUse().getItem() instanceof ItemSpear || player.getItemInUse().getItem() instanceof ItemHammer || player.getItemInUse().getItem() instanceof ItemGrappHook))
+          {
+               int i = player.getItemInUseDuration();
+               float f1 = (float) i / 20.0F;
+               
+               if (f1 > 1.0F)
+               {
+                    f1 = 1.0F;
+               }
+               else
+               {
+                    f1 *= f1;
+               }
+               
+               event.newfov *= 1.0F - f1 * 0.15F;
+          }
+     }
 }
